Exported the remainder of the enums to Utilities.Enums
