+++ RelhaxModpack/RelhaxModpack/App.xaml.cs
@@ -8,6 +8,7 @@ using RelhaxModpack.Patches;
 using RelhaxModpack.Utilities;
 using RelhaxModpack.Windows;
 using RelhaxModpack.Xml;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack
 {
