+++ RelhaxModpack/RelhaxModpack/Atlases/Packing/ImagePacker.cs
@@ -5,6 +5,7 @@ using System.Drawing.Imaging;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Atlases.Packing
 {
