+++ RelhaxModpack/RelhaxModpack/Common/Translations.cs
@@ -10,45 +10,10 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Documents;
 using System.Windows.Media;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack
 {
-    /// <summary>
-    /// List of all supported Languages in the modpack
-    /// </summary>
-    public enum Languages
-    {
-        /// <summary>
-        /// The English language
-        /// </summary>
-        English,
-
-        /// <summary>
-        /// The German language
-        /// </summary>
-        German,
-
-        /// <summary>
-        /// The Polish language
-        /// </summary>
-        Polish,
-
-        /// <summary>
-        /// The French language
-        /// </summary>
-        French,
-
-        /// <summary>
-        /// The Spanish language
-        /// </summary>
-        Spanish,
-
-        /// <summary>
-        /// The Russian language
-        /// </summary>
-        Russian
-    };
-
     /// <summary>
     /// Handles all localization for the application User Interface
     /// </summary>
