+++ RelhaxModpack/RelhaxModpack/Database/DownloadInstructions.cs
@@ -1,17 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Database
 {
-    public enum DownloadTypes
-    {
-        StaticLink,
-        WgMods,
-        WebScrape
-    }
     public class DownloadInstructions
     {
         public string InstructionsVersion { get; set; }
