+++ RelhaxModpack/RelhaxModpack/Patches/PatchUtils.cs
@@ -16,27 +16,6 @@ using RelhaxModpack.Utilities.Enums;
 namespace RelhaxModpack.Patches
 {
     /// <summary>
-    /// Exit codes during a patch operation
-    /// </summary>
-    public enum PatchExitCode
-    {
-        /// <summary>
-        /// The patch operation succeeded
-        /// </summary>
-        Success = 0,
-
-        /// <summary>
-        /// The patch operation succeeded, but with warnings
-        /// </summary>
-        Warning = -1,
-
-        /// <summary>
-        /// The patch operation did not succeeded
-        /// </summary>
-        Error = -2
-    }
-
-    /// <summary>
     /// A class for handling patch operations
     /// </summary>
     public static class PatchUtils
