+++ RelhaxModpack/RelhaxModpack/Patches/Regression.cs
@@ -5,35 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 using System.IO;
 using RelhaxModpack.Utilities;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Patches
 {
-    /// <summary>
-    /// The types of patch regression tests that can be performed
-    /// </summary>
-    public enum PatchRegressionTypes
-    {
-        /// <summary>
-        /// Json regression (standard, non-XVM)
-        /// </summary>
-        json,
-
-        /// <summary>
-        /// Xml regression
-        /// </summary>
-        xml,
-
-        /// <summary>
-        /// Regex regression
-        /// </summary>
-        regex,
-
-        /// <summary>
-        /// Json regression (XVM style)
-        /// </summary>
-        followPath
-    }
-
     /// <summary>
     /// Represents a patch operation with a description and desired assertion condition
     /// </summary>
