+++ RelhaxModpack/RelhaxModpack/RelhaxModpack.csproj
@@ -173,10 +173,27 @@
     <Compile Include="UI\UiUtils.cs" />
     <Compile Include="UI\WindowColorset.cs" />
     <Compile Include="UI\ZoomBorder.cs" />
+    <Compile Include="Utilities\Enums\ApplicationMode.cs" />
+    <Compile Include="Utilities\Enums\ApplyBehavior.cs" />
+    <Compile Include="Utilities\Enums\DatabaseVersions.cs" />
     <Compile Include="Utilities\Enums\DatabaseXmlVersion.cs" />
+    <Compile Include="Utilities\Enums\DownloadTypes.cs" />
+    <Compile Include="Utilities\Enums\FailCode.cs" />
     <Compile Include="Utilities\Enums\KnownFolder.cs" />
     <Compile Include="Utilities\Enums\KnownFolderFlags.cs" />
+    <Compile Include="Utilities\Enums\Languages.cs" />
+    <Compile Include="Utilities\Enums\LogFiles.cs" />
+    <Compile Include="Utilities\Enums\LogLevel.cs" />
+    <Compile Include="Utilities\Enums\LogOptions.cs" />
+    <Compile Include="Utilities\Enums\MediaType.cs" />
+    <Compile Include="Utilities\Enums\PatchExitCode.cs" />
+    <Compile Include="Utilities\Enums\PatchRegressionTypes.cs" />
     <Compile Include="Utilities\Enums\ReplacementTypes.cs" />
+    <Compile Include="Utilities\Enums\SelectionTypes.cs" />
+    <Compile Include="Utilities\Enums\SelectionView.cs" />
+    <Compile Include="Utilities\Enums\UIThemes.cs" />
+    <Compile Include="Utilities\Enums\UninstallModes.cs" />
+    <Compile Include="Utilities\Enums\UpdateTypes.cs" />
     <Compile Include="Utilities\Enums\XmlLoadType.cs" />
     <Compile Include="Utilities\FileUtils.cs" />
     <Compile Include="Utilities\FtpUtils.cs" />
