+++ RelhaxModpack/RelhaxModpack/UI/UiUtils.cs
@@ -5,6 +5,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Threading;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.UI
 {
