+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/DatabaseVersions.cs
@@ -1,0 +1,23 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// Database distribution levels
+    /// </summary>
+    public enum DatabaseVersions
+    {
+        /// <summary>
+        /// The stable public database
+        /// </summary>
+        Stable,
+
+        /// <summary>
+        /// The unstable public beta database
+        /// </summary>
+        Beta,
+
+        /// <summary>
+        /// The unstable private testing database.
+        /// </summary>
+        Test
+    }
+}
