+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/DownloadTypes.cs
@@ -1,0 +1,9 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    public enum DownloadTypes
+    {
+        StaticLink,
+        WgMods,
+        WebScrape
+    }
+}
