+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/Languages.cs
@@ -1,0 +1,38 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// List of all supported Languages in the modpack
+    /// </summary>
+    public enum Languages
+    {
+        /// <summary>
+        /// The English language
+        /// </summary>
+        English,
+
+        /// <summary>
+        /// The German language
+        /// </summary>
+        German,
+
+        /// <summary>
+        /// The Polish language
+        /// </summary>
+        Polish,
+
+        /// <summary>
+        /// The French language
+        /// </summary>
+        French,
+
+        /// <summary>
+        /// The Spanish language
+        /// </summary>
+        Spanish,
+
+        /// <summary>
+        /// The Russian language
+        /// </summary>
+        Russian
+    }
+}
