+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/LogOptions.cs
@@ -1,0 +1,10 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    public enum LogOptions
+    {
+        None, 
+        MethodName, 
+        ClassName, 
+        MethodAndClassName
+    }
+}
