+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/SelectionView.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// The type of selection view for how to display the selection tree
+    /// </summary>
+    public enum SelectionView
+    {
+        /// <summary>
+        /// OMC style
+        /// </summary>
+        Legacy = 1,
+
+        /// <summary>
+        /// Default WPF V2 style
+        /// </summary>
+        DefaultV2 = 0
+    }
+}
