+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/UIThemes.cs
@@ -1,0 +1,23 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// The list of different types of themes in the application
+    /// </summary>
+    public enum UIThemes
+    {
+        /// <summary>
+        /// The default application theme
+        /// </summary>
+        Default,
+
+        /// <summary>
+        /// The dark application theme
+        /// </summary>
+        Dark,
+
+        /// <summary>
+        /// A custom application theme from the UISettings.xml
+        /// </summary>
+        Custom
+    }
+}
