+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/UninstallModes.cs
@@ -1,0 +1,18 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    /// <summary>
+    /// The types on uninstall mods the application supports
+    /// </summary>
+    public enum UninstallModes
+    {
+        /// <summary>
+        /// Default uninstall method to uninstall all modifications done by the application
+        /// </summary>
+        Default = 0,
+
+        /// <summary>
+        /// Quick uninstall method to remove 
+        /// </summary>
+        Quick = 1
+    }
+}
