+++ RelhaxModpack/RelhaxModpack/Utilities/Enums/UpdateTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace RelhaxModpack.Utilities.Enums
+{
+    public enum UpdateTypes
+    {
+        zip,
+        wotmod
+    }
+}
