+++ RelhaxModpack/RelhaxModpack/Windows/EditorAddRemove.xaml.cs
@@ -13,6 +13,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using RelhaxModpack.Database;
 using RelhaxModpack.UI;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Windows
 {
