+++ RelhaxModpack/RelhaxModpack/Xml/XmlBinaryHandler.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.IO;
 using System.Xml;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxModpack.Xml
 {
