+++ RelhaxModpack/RelhaxUnitTests/UnitTestHelper.cs
@@ -8,6 +8,7 @@ using RelhaxModpack;
 using System.IO;
 using System.Diagnostics;
 using System.Reflection;
+using RelhaxModpack.Utilities.Enums;
 
 namespace RelhaxUnitTests
 {
